refactor(lock): split manager.go by the concerns it already names
manager.go was 3540 lines carrying nine banner-delimited sections. Each
section moves to the file for its concern: lease wrappers to leases.go,
POSIX split/merge/upgrade to a new byterange.go, the caching-break
operations and callback registration to a new break.go, delegation CRUD
to delegation.go, grace-period delegation to grace.go, and
connection/cleanup to connection.go.

Pure code motion — every declaration and every non-import line is
preserved verbatim; only import blocks differ.
