+++ pkg/metadata/lock/manager.go
@@ -1,11 +1,8 @@
 package lock
 
 import (
-	"cmp"
 	"context"
-	"errors"
 	"fmt"
-	"slices"
 	"strconv"
 	"strings"
 	"sync"
@@ -1376,2165 +1373,4 @@ func sessionIDFromOwnerID(ownerID string) (uint64, bool) {
 		return 0, false
 	}
 	return sid, true
-}
-
-// ============================================================================
-// Lease Operations (implementations in leases.go and reclaim.go)
-// ============================================================================
-
-// RequestLease requests a new or upgraded lease on a file or directory.
-func (lm *Manager) RequestLease(ctx context.Context, fileHandle FileHandle, leaseKey [16]byte,
-	parentLeaseKey [16]byte, ownerID string, clientID string, shareName string,
-	requestedState uint32, isDirectory bool) (grantedState uint32, epoch uint16, err error) {
-	return lm.requestLeaseImpl(ctx, leaseRequest{
-		fileHandle:            fileHandle,
-		leaseKey:              leaseKey,
-		parentLeaseKey:        parentLeaseKey,
-		ownerID:               ownerID,
-		clientID:              clientID,
-		shareName:             shareName,
-		state:                 requestedState,
-		isDirectory:           isDirectory,
-		isTraditionalOplock:   false,
-		suppressConflictBreak: false,
-	})
-}
-
-// RequestLeaseAsOplock is the traditional-oplock variant of RequestLease.
-// The SMB adapter calls this when a CREATE arrives with a non-Lease
-// OplockLevel (LEVEL_II / Exclusive / Batch); the new record is tagged
-// `IsTraditionalOplock=true` so subsequent grants observe the cross-tier
-// rules described in `bestGrantableState`. All other parameters and
-// semantics match `RequestLease`.
-//
-// Reference: MS-SMB2 §3.3.5.9 / Samba `source3/smbd/open.c::grant_fsp_oplock_type`.
-func (lm *Manager) RequestLeaseAsOplock(ctx context.Context, fileHandle FileHandle, leaseKey [16]byte,
-	parentLeaseKey [16]byte, ownerID string, clientID string, shareName string,
-	requestedState uint32, isDirectory bool) (grantedState uint32, epoch uint16, err error) {
-	return lm.requestLeaseImpl(ctx, leaseRequest{
-		fileHandle:            fileHandle,
-		leaseKey:              leaseKey,
-		parentLeaseKey:        parentLeaseKey,
-		ownerID:               ownerID,
-		clientID:              clientID,
-		shareName:             shareName,
-		state:                 requestedState,
-		isDirectory:           isDirectory,
-		isTraditionalOplock:   true,
-		suppressConflictBreak: false,
-	})
-}
-
-// RequestLeaseStatOpen is the stat-open variant of RequestLease. The SMB
-// adapter calls this when a CREATE's DesiredAccess is stat-open-only
-// (FILE_READ_ATTRIBUTES / FILE_WRITE_ATTRIBUTES / READ_CONTROL / SYNCHRONIZE) and
-// the disposition is non-destructive. The grant proceeds normally except that
-// a cross-key conflict with an existing holder MUST NOT dispatch a break: a
-// stat-opener caches attributes alongside existing holders without forcing
-// them to drop their caches. It instead receives the best state it can coexist
-// with (`bestGrantableState`).
-//
-// Reference: MS-SMB2 §3.3.5.9.8 / Samba `is_lease_stat_open`
-// (source3/smbd/open.c). Closes the timing-dependent spurious break that
-// smb2.lease.statopen4 CHECK_NO_BREAK observes (#751).
-func (lm *Manager) RequestLeaseStatOpen(ctx context.Context, fileHandle FileHandle, leaseKey [16]byte,
-	parentLeaseKey [16]byte, ownerID string, clientID string, shareName string,
-	requestedState uint32, isDirectory bool) (grantedState uint32, epoch uint16, err error) {
-	return lm.requestLeaseImpl(ctx, leaseRequest{
-		fileHandle:            fileHandle,
-		leaseKey:              leaseKey,
-		parentLeaseKey:        parentLeaseKey,
-		ownerID:               ownerID,
-		clientID:              clientID,
-		shareName:             shareName,
-		state:                 requestedState,
-		isDirectory:           isDirectory,
-		isTraditionalOplock:   false,
-		suppressConflictBreak: true,
-	})
-}
-
-// AcknowledgeLeaseBreak processes a client's lease break acknowledgment.
-func (lm *Manager) AcknowledgeLeaseBreak(ctx context.Context, leaseKey [16]byte,
-	acknowledgedState uint32, epoch uint16) error {
-	return lm.acknowledgeLeaseBreakImpl(ctx, leaseKey, acknowledgedState, epoch)
-}
-
-// ReleaseLease releases all lease state for the given lease key.
-func (lm *Manager) ReleaseLease(ctx context.Context, leaseKey [16]byte) error {
-	return lm.releaseLeaseImpl(ctx, leaseKey)
-}
-
-// ReclaimLease reclaims a lease during grace period. clientID must match the
-// lease owner recorded on the persisted record (lease-stealing guard); pass ""
-// to skip the owner check. This is the LockManager-interface entrypoint used by
-// SMB, which has no RPCSEC_GSS principal to verify.
-func (lm *Manager) ReclaimLease(ctx context.Context, leaseKey [16]byte,
-	requestedState uint32, isDirectory bool, clientID string) (*UnifiedLock, error) {
-	return lm.reclaimLeaseImpl(ctx, leaseKey, requestedState, isDirectory, clientID, "")
-}
-
-// ReclaimLeaseWithPrincipal is the NFSv4 reclaim path that additionally
-// verifies the incoming RPCSEC_GSS / AUTH_SYS principal matches the one
-// recorded in the V4ClientRecoveryRecord for clientID. The principal check
-// runs only when a ClientRecoveryStore is wired (SetClientRecoveryStore) and
-// incomingPrincipal is non-empty; an empty principal skips the check. Returns
-// a lock-not-found error when the clientID or principal does not match.
-func (lm *Manager) ReclaimLeaseWithPrincipal(ctx context.Context, leaseKey [16]byte,
-	requestedState uint32, isDirectory bool, clientID string, incomingPrincipal string) (*UnifiedLock, error) {
-	return lm.reclaimLeaseImpl(ctx, leaseKey, requestedState, isDirectory, clientID, incomingPrincipal)
-}
-
-// ReleaseLeaseForHandle removes lease records matching leaseKey from a
-// single handleKey bucket only. See releaseLeaseForHandleImpl for details.
-func (lm *Manager) ReleaseLeaseForHandle(ctx context.Context, handleKey string, leaseKey [16]byte) error {
-	return lm.releaseLeaseForHandleImpl(ctx, handleKey, leaseKey)
-}
-
-// GetLeaseState returns the current state and epoch for a lease key.
-func (lm *Manager) GetLeaseState(ctx context.Context, leaseKey [16]byte) (state uint32, epoch uint16, found bool) {
-	return lm.getLeaseStateImpl(ctx, leaseKey)
-}
-
-// SetLeaseEpoch sets the epoch on an existing lease identified by leaseKey.
-// Per MS-SMB2 3.3.5.9: For V2 leases, the server should track the client's
-// epoch from the RqLs create context. SetLeaseEpoch is called after RequestLease
-// to initialize the epoch to the client's requested value.
-// Returns false if no lease was found with the given key.
-func (lm *Manager) SetLeaseEpoch(leaseKey [16]byte, epoch uint16) bool {
-	lm.mu.Lock()
-	defer lm.mu.Unlock()
-
-	// Update every lease record matching leaseKey, not just the first found.
-	// Stale records from prior tests (same LEASE1 constant across smbtorture
-	// tests) or from multiple opens by different clients can coexist in
-	// lm.unifiedLocks under different handleKey buckets. findLeaseByKey's
-	// map-iteration order is non-deterministic, so scoping to the first
-	// match can miss the lease that RequestLease just granted — leaving it
-	// at Epoch=1 (createAndGrantLease default) while the response to the
-	// client carries the higher requested epoch. Subsequent break
-	// notifications then dispatch with Epoch=2 instead of requestedEpoch+2,
-	// regressing smbtorture V2 tests (break_twice, breaking*, v2_breaking3).
-	// Two passes so all matching records converge to the SAME epoch. A
-	// per-record `if epoch >= lock.Lease.Epoch` guard lets sibling records that
-	// start at different epochs diverge (e.g. one at 5 stays 5 while another at
-	// 1 moves to 4), and GetLeaseState / break dispatch then read whichever
-	// map-order surfaces first — a non-deterministic NewEpoch in break
-	// notifications (the break_twice / v2_breaking3 regression class). Compute
-	// the max of the requested epoch and every matching record's current epoch,
-	// then assign that single max to all of them so the lease has one epoch.
-	//
-	// Probes only the buckets leaseKeyIndex names for leaseKey; the ones it
-	// omits hold no record for the key and could never have matched.
-	target := epoch
-	var matches []*UnifiedLock
-	for handleKey := range lm.leaseKeyIndex[leaseKey] {
-		for _, lock := range lm.unifiedLocks[handleKey] {
-			if lock.Lease == nil || lock.Lease.LeaseKey != leaseKey {
-				continue
-			}
-			if lock.Lease.Epoch > target {
-				target = lock.Lease.Epoch
-			}
-			matches = append(matches, lock)
-		}
-	}
-	for _, lock := range matches {
-		lock.Lease.Epoch = target
-		// Persist the new epoch like every other lease state-change path.
-		// Without this, RestoreLocks rebuilds the lease at the stale grant-time
-		// epoch after a restart and a later break dispatches a NewEpoch below
-		// the client's last-seen value, violating MS-SMB2 §2.2.14.2.11 /
-		// §3.3.4.7 epoch monotonicity.
-		lm.persistUnifiedLockLocked(lock)
-	}
-	return len(matches) > 0
-}
-
-// ============================================================================
-// POSIX Lock Splitting
-// ============================================================================
-
-// SplitLock splits an existing lock when a portion is unlocked.
-//
-// POSIX semantics require that unlocking a portion of a locked range results in:
-//   - 0 locks: if the unlock range covers the entire lock
-//   - 1 lock: if the unlock range covers the start or end
-//   - 2 locks: if the unlock range is in the middle (creates a "hole")
-//
-// Parameters:
-//   - existing: The lock to split
-//   - unlockOffset: Starting byte offset of the unlock range
-//   - unlockLength: Number of bytes to unlock (0 = to EOF)
-//
-// Returns:
-//   - []UnifiedLock: The resulting locks after the split (0, 1, or 2 locks)
-//
-// Examples:
-//   - Lock [0-100], Unlock [0-100] -> [] (exact match)
-//   - Lock [0-100], Unlock [0-50] -> [[50-100]] (unlock at start)
-//   - Lock [0-100], Unlock [50-100] -> [[0-50]] (unlock at end)
-//   - Lock [0-100], Unlock [25-75] -> [[0-25], [75-100]] (unlock in middle)
-func SplitLock(existing *UnifiedLock, unlockOffset, unlockLength uint64) []*UnifiedLock {
-	// Check if ranges overlap at all
-	if !RangesOverlap(existing.Offset, existing.Length, unlockOffset, unlockLength) {
-		// No overlap - return existing lock unchanged
-		return []*UnifiedLock{existing.Clone()}
-	}
-
-	// Calculate lock end. End() returns maxUint64 for unbounded (Length==0)
-	// locks, which is exactly the "treat as very large" value the coverage
-	// arithmetic below needs.
-	lockEnd := existing.End()
-
-	// Calculate unlock end. unlockOffset/unlockLength are wire-controlled, so
-	// the sum can wrap uint64 (e.g. offset near 2^64 + a large length); a wrap
-	// would corrupt the coverage checks below (fabricate or drop a fragment).
-	// Clamp to maxUint64 on overflow — a range that reaches past 2^64 covers to
-	// EOF for our purposes, same as an unbounded unlock.
-	unlockEnd := unlockOffset + unlockLength
-	if unlockLength == 0 || unlockEnd < unlockOffset {
-		// Unbounded unlock (length 0 = to EOF) or arithmetic overflow.
-		unlockEnd = ^uint64(0)
-	}
-
-	// Check for exact match or complete coverage
-	if unlockOffset <= existing.Offset && unlockEnd >= lockEnd {
-		// Unlock completely covers the lock - remove it
-		return []*UnifiedLock{}
-	}
-
-	var result []*UnifiedLock
-
-	// Each fragment is a distinct lock and MUST get a fresh ID. Clone() copies
-	// the original's ID verbatim, so two fragments would otherwise share one
-	// persist identity — the second PutLock would overwrite the first (the
-	// store is keyed by ID), silently losing one byte-range across a restart.
-	// Check if there's a portion before the unlock range
-	if unlockOffset > existing.Offset {
-		beforeLock := existing.Clone()
-		beforeLock.ID = uuid.New().String()
-		beforeLock.Length = unlockOffset - existing.Offset
-		result = append(result, beforeLock)
-	}
-
-	// Check if there's a portion after the unlock range
-	if unlockEnd < lockEnd {
-		afterLock := existing.Clone()
-		afterLock.ID = uuid.New().String()
-		afterLock.Offset = unlockEnd
-		if existing.Length == 0 {
-			// Original was unbounded, after portion is also unbounded
-			afterLock.Length = 0
-		} else {
-			afterLock.Length = lockEnd - unlockEnd
-		}
-		result = append(result, afterLock)
-	}
-
-	return result
-}
-
-// ============================================================================
-// Lock Merging
-// ============================================================================
-
-// MergeLocks coalesces adjacent or overlapping locks from the same owner.
-//
-// This is used when upgrading or extending locks to avoid fragmentation.
-// Only locks with the same owner, type, and file handle can be merged.
-//
-// Parameters:
-//   - locks: Slice of locks to potentially merge
-//
-// Returns:
-//   - []UnifiedLock: Merged locks (may have fewer elements than input)
-func MergeLocks(locks []*UnifiedLock) []*UnifiedLock {
-	if len(locks) == 0 {
-		return nil
-	}
-	if len(locks) == 1 {
-		return []*UnifiedLock{locks[0].Clone()}
-	}
-
-	// Group locks by owner+type+filehandle
-	type groupKey struct {
-		ownerID    string
-		lockType   LockType
-		fileHandle string
-	}
-
-	groups := make(map[groupKey][]*UnifiedLock)
-	for _, lock := range locks {
-		key := groupKey{
-			ownerID:    lock.Owner.OwnerID,
-			lockType:   lock.Type,
-			fileHandle: string(lock.FileHandle),
-		}
-		groups[key] = append(groups[key], lock)
-	}
-
-	var result []*UnifiedLock
-
-	for _, group := range groups {
-		merged := mergeRanges(group)
-		result = append(result, merged...)
-	}
-
-	return result
-}
-
-// mergeRanges merges locks that have the same owner/type/file.
-// It combines overlapping or adjacent ranges into single locks.
-func mergeRanges(locks []*UnifiedLock) []*UnifiedLock {
-	if len(locks) == 0 {
-		return nil
-	}
-	if len(locks) == 1 {
-		return []*UnifiedLock{locks[0].Clone()}
-	}
-
-	// Sort by offset
-	sorted := make([]*UnifiedLock, len(locks))
-	for i, l := range locks {
-		sorted[i] = l.Clone()
-	}
-	slices.SortFunc(sorted, func(a, b *UnifiedLock) int {
-		return cmp.Compare(a.Offset, b.Offset)
-	})
-
-	var result []*UnifiedLock
-	current := sorted[0]
-
-	for i := 1; i < len(sorted); i++ {
-		next := sorted[i]
-
-		// Check if current and next can be merged
-		if canMerge(current, next) {
-			// Merge into current
-			current = mergeTwoLocks(current, next)
-		} else {
-			// Can't merge - finalize current and move to next
-			result = append(result, current)
-			current = next
-		}
-	}
-
-	// Don't forget the last one
-	result = append(result, current)
-
-	return result
-}
-
-// canMerge checks if two locks can be merged (adjacent or overlapping).
-func canMerge(a, b *UnifiedLock) bool {
-	// Must be same owner, type, and file (assumed by caller grouping)
-
-	// Handle unbounded locks
-	if a.Length == 0 {
-		// a is unbounded - can merge with anything at or after a.Offset
-		return b.Offset >= a.Offset
-	}
-	if b.Length == 0 {
-		// b is unbounded - can merge if a overlaps or is adjacent to b.Offset
-		return a.End() >= b.Offset
-	}
-
-	// Both bounded - check if adjacent or overlapping
-	aEnd := a.End()
-	return aEnd >= b.Offset // Adjacent (aEnd == b.Offset) or overlapping
-}
-
-// mergeTwoLocks combines two locks into one.
-func mergeTwoLocks(a, b *UnifiedLock) *UnifiedLock {
-	result := a.Clone()
-
-	// Start is the minimum offset
-	result.Offset = min(a.Offset, b.Offset)
-
-	// Handle unbounded locks
-	if a.Length == 0 || b.Length == 0 {
-		result.Length = 0 // Result is unbounded
-		return result
-	}
-
-	// Both bounded - end is the maximum
-	maxEnd := max(a.End(), b.End())
-
-	result.Length = maxEnd - result.Offset
-	return result
-}
-
-// ============================================================================
-// Atomic Lock Upgrade
-// ============================================================================
-
-// UpgradeLock atomically converts a shared lock to exclusive if no other readers exist.
-//
-// This implements the user decision: "Lock upgrade: Atomic upgrade supported
-// (read -> write if no other readers)".
-//
-// Steps:
-//  1. Find existing shared lock owned by `owner` covering the range
-//  2. Check if any OTHER owners hold shared locks on overlapping range
-//  3. If other readers exist: return ErrLockConflict
-//  4. If no other readers: atomically change lock type to Exclusive
-//
-// Parameters:
-//   - handleKey: The file handle key
-//   - owner: The lock owner requesting the upgrade
-//   - offset: Starting byte offset of the range to upgrade
-//   - length: Number of bytes (0 = to EOF)
-//
-// Returns:
-//   - *UnifiedLock: The upgraded lock on success
-//   - error: ErrLockConflict if other readers exist, ErrLockNotFound if no lock to upgrade
-//
-// PRECONDITION — whole-lock upgrade only. Step 3 flips the ENTIRE matched
-// shared lock to exclusive, not just the [offset,length) sub-range, and that
-// whole-range promotion is what gets persisted. This is correct ONLY because
-// every caller upgrades a lock at exactly the range it was granted at: the
-// NFSv4 LOCK upgrade path promotes the lock-owner's existing lock as a whole,
-// never a strict sub-range of a larger shared lock. If a caller ever passes a
-// sub-range of a wider shared lock, the surrounding bytes would be silently
-// promoted to exclusive (and persisted that way) — split the matched lock into
-// the upgraded sub-range plus shared remainder(s) before changing this. There
-// are currently no production callers (NFSv4/NLM upgrade is not yet wired); the
-// invariant is enforced by TestUpgradeLock_WholeLockOnly.
-func (lm *Manager) UpgradeLock(handleKey string, owner LockOwner, offset, length uint64) (*UnifiedLock, error) {
-	lm.mu.Lock()
-	defer lm.mu.Unlock()
-
-	unifiedLocks := lm.getUnifiedLocksLocked(handleKey)
-
-	// Step 1: Find existing shared lock owned by this owner covering the range
-	var ownLock *UnifiedLock
-	var ownLockIndex = -1
-
-	for i, lock := range unifiedLocks {
-		if lock.Owner.OwnerID == owner.OwnerID &&
-			lock.Type == LockTypeShared &&
-			lock.Overlaps(offset, length) {
-			// Found our shared lock
-			ownLock = lock
-			ownLockIndex = i
-			break
-		}
-	}
-
-	if ownLock == nil {
-		// Check if we already have an exclusive lock (no-op case)
-		for _, lock := range unifiedLocks {
-			if lock.Owner.OwnerID == owner.OwnerID &&
-				lock.Type == LockTypeExclusive &&
-				lock.Overlaps(offset, length) {
-				// Already exclusive - return it as-is
-				return lock.Clone(), nil
-			}
-		}
-		return nil, NewLockNotFoundError("")
-	}
-
-	// Step 2: Check if any OTHER owners hold shared locks on overlapping range
-	for _, lock := range unifiedLocks {
-		if lock.Owner.OwnerID == owner.OwnerID {
-			continue // Skip our own locks
-		}
-		if lock.Overlaps(offset, length) {
-			// Another owner has a lock on this range - cannot upgrade
-			return nil, NewLockConflictError("", &UnifiedLockConflict{
-				Lock:   lock,
-				Reason: "other reader exists on range",
-			})
-		}
-	}
-
-	// Step 3: Atomically upgrade the lock
-	unifiedLocks[ownLockIndex].Type = LockTypeExclusive
-
-	// Persist the upgraded type under lm.mu so the change survives a restart.
-	// Without this the in-memory lock reverted to shared on restart and a
-	// reader could be wrongly granted against an intended-exclusive lock (R3-3).
-	lm.persistUnifiedLockLocked(unifiedLocks[ownLockIndex])
-
-	return unifiedLocks[ownLockIndex].Clone(), nil
-}
-
-// getUnifiedLocksLocked returns unified locks for a file (must hold lm.mu).
-func (lm *Manager) getUnifiedLocksLocked(handleKey string) []*UnifiedLock {
-	return lm.unifiedLocks[handleKey]
-}
-
-// AddUnifiedLock adds a unified lock to the storage.
-//
-// Checks for conflicts using the ConflictsWith method which handles all 4
-// conflict cases: access modes, oplock-oplock, oplock-byterange, byterange-byterange.
-func (lm *Manager) AddUnifiedLock(handleKey string, lock *UnifiedLock) error {
-	lm.mu.Lock()
-	defer lm.mu.Unlock()
-
-	existing := lm.unifiedLocks[handleKey]
-
-	// Check for conflicts with existing locks using ConflictsWith
-	for _, el := range existing {
-		if lock.ConflictsWith(el) {
-			return NewLockConflictError("", &UnifiedLockConflict{
-				Lock:   el,
-				Reason: "lock conflict",
-			})
-		}
-	}
-
-	// Cross-protocol: an overlapping SMB byte-range lock must also block this
-	// NLM/NFSv4 lock (area-5 H-3 / xproto H1). Skip for whole-file leases and
-	// delegations, which are not byte-range locks.
-	if !lock.IsLease() && !lock.IsDelegation() {
-		smbLocks := lm.locks[handleKey]
-		for i := range smbLocks {
-			if fileLockConflictsWithUnified(&smbLocks[i], lock) {
-				return NewLockConflictError("", unifiedConflictFromFileLock(&smbLocks[i]))
-			}
-		}
-	}
-
-	// Check if this exact lock already exists (same owner, offset, length)
-	// If so, update it (allows changing lock type)
-	for i, el := range existing {
-		if el.Owner.OwnerID == lock.Owner.OwnerID &&
-			el.Offset == lock.Offset &&
-			el.Length == lock.Length {
-			// Update existing lock in place
-			existing[i].Type = lock.Type
-			existing[i].AcquiredAt = time.Now()
-			lm.persistUnifiedLockLocked(existing[i])
-			return nil
-		}
-	}
-
-	// Set acquisition time if not set
-	if lock.AcquiredAt.IsZero() {
-		lock.AcquiredAt = time.Now()
-	}
-
-	// Add new lock
-	lm.unifiedLocks[handleKey] = append(existing, lock)
-	lm.indexAddLockLocked(handleKey, lock)
-	lm.persistUnifiedLockLocked(lock)
-	return nil
-}
-
-// TestUnifiedLock previews whether a prospective NLM/NFSv4 byte-range lock would
-// conflict, checking BOTH the unified lock map and the SMB byte-range map
-// (lm.locks). Returns the conflicting lock, or nil if grantable. This keeps the
-// NLM/NFSv4 TEST/LOCKT preview consistent with AddUnifiedLock enforcement: a
-// preview that only scanned unifiedLocks would report a range grantable while
-// an overlapping SMB byte-range lock in lm.locks would deny the acquire
-// (area-5 H-3 / xproto H1).
-func (lm *Manager) TestUnifiedLock(handleKey string, want *UnifiedLock) *UnifiedLockConflict {
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-
-	for _, el := range lm.unifiedLocks[handleKey] {
-		if want.ConflictsWith(el) {
-			return &UnifiedLockConflict{Lock: el, Reason: "lock conflict"}
-		}
-	}
-
-	if !want.IsLease() && !want.IsDelegation() {
-		smbLocks := lm.locks[handleKey]
-		for i := range smbLocks {
-			if fileLockConflictsWithUnified(&smbLocks[i], want) {
-				return unifiedConflictFromFileLock(&smbLocks[i])
-			}
-		}
-	}
-
-	return nil
-}
-
-// RemoveUnifiedLock removes a unified lock using POSIX splitting semantics.
-func (lm *Manager) RemoveUnifiedLock(handleKey string, owner LockOwner, offset, length uint64) error {
-	lm.mu.Lock()
-	defer lm.mu.Unlock()
-
-	existing := lm.unifiedLocks[handleKey]
-	if len(existing) == 0 {
-		return NewLockNotFoundError("")
-	}
-
-	var newLocks []*UnifiedLock
-	found := false
-
-	for _, lock := range existing {
-		if lock.Owner.OwnerID != owner.OwnerID {
-			// Not our lock - keep it
-			newLocks = append(newLocks, lock)
-			continue
-		}
-
-		// Our lock - check if it overlaps with the unlock range
-		if !lock.Overlaps(offset, length) {
-			// Doesn't overlap - keep it unchanged
-			newLocks = append(newLocks, lock)
-			continue
-		}
-
-		// Overlaps - split the lock. Delete the original record, then persist
-		// each fragment (each carries a fresh UUID from SplitLock). Done under
-		// lm.mu so the store sees delete-then-puts in mutation order.
-		found = true
-		lm.deleteUnifiedLockLocked(lock)
-		splitResult := SplitLock(lock, offset, length)
-		for _, frag := range splitResult {
-			lm.persistUnifiedLockLocked(frag)
-		}
-		newLocks = append(newLocks, splitResult...)
-	}
-
-	if !found {
-		return NewLockNotFoundError("")
-	}
-
-	// Update or clean up
-	if len(newLocks) == 0 {
-		delete(lm.unifiedLocks, handleKey)
-	} else {
-		lm.unifiedLocks[handleKey] = newLocks
-	}
-	lm.reindexHandleLocked(handleKey, existing)
-
-	return nil
-}
-
-// ListUnifiedLocks returns all unified locks on a file.
-func (lm *Manager) ListUnifiedLocks(handleKey string) []*UnifiedLock {
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-
-	existing := lm.unifiedLocks[handleKey]
-	if len(existing) == 0 {
-		return nil
-	}
-
-	// Return a copy to avoid race conditions
-	result := make([]*UnifiedLock, len(existing))
-	for i, el := range existing {
-		result[i] = el.Clone()
-	}
-	return result
-}
-
-// RemoveFileUnifiedLocks removes all unified locks, delegations, and break
-// wait channels for a file.
-func (lm *Manager) RemoveFileUnifiedLocks(handleKey string) {
-	lm.mu.Lock()
-	old := lm.unifiedLocks[handleKey]
-	delete(lm.unifiedLocks, handleKey)
-	lm.reindexHandleLocked(handleKey, old)
-	delete(lm.breakWaitChans, handleKey)
-	lm.mu.Unlock()
-}
-
-// GetUnifiedLock retrieves a specific unified lock by owner and range.
-//
-// Returns the matching lock or ErrLockNotFound if no matching lock exists.
-func (lm *Manager) GetUnifiedLock(handleKey string, owner LockOwner, offset, length uint64) (*UnifiedLock, error) {
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-
-	for _, lock := range lm.unifiedLocks[handleKey] {
-		if lock.Owner.OwnerID == owner.OwnerID &&
-			lock.Offset == offset &&
-			lock.Length == length {
-			return lock.Clone(), nil
-		}
-	}
-
-	return nil, NewLockNotFoundError("")
-}
-
-// CheckAndBreakOpLocksForWrite checks and initiates breaks for oplocks that
-// conflict with a write operation. Backward-compatible wrapper for CheckAndBreakCachingForWrite.
-func (lm *Manager) CheckAndBreakOpLocksForWrite(handleKey string, excludeOwner *LockOwner) error {
-	return lm.CheckAndBreakCachingForWrite(handleKey, excludeOwner)
-}
-
-// CheckAndBreakOpLocksForRead checks and initiates breaks for oplocks that
-// conflict with a read operation. Backward-compatible wrapper for CheckAndBreakCachingForRead.
-func (lm *Manager) CheckAndBreakOpLocksForRead(handleKey string, excludeOwner *LockOwner) error {
-	return lm.CheckAndBreakCachingForRead(handleKey, excludeOwner)
-}
-
-// CheckAndBreakOpLocksForDelete checks and initiates breaks for all oplocks
-// on a file being deleted. Backward-compatible wrapper for CheckAndBreakCachingForDelete.
-func (lm *Manager) CheckAndBreakOpLocksForDelete(handleKey string, excludeOwner *LockOwner) error {
-	return lm.CheckAndBreakCachingForDelete(handleKey, excludeOwner)
-}
-
-// ============================================================================
-// Unified Caching Break Operations
-// ============================================================================
-
-// CheckAndBreakCachingForWrite breaks all leases AND all delegations.
-// Used for cross-protocol writes (e.g., NFS write breaking SMB leases).
-func (lm *Manager) CheckAndBreakCachingForWrite(handleKey string, excludeOwner *LockOwner) error {
-	if err := lm.breakOpLocks(handleKey, excludeOwner, LeaseStateNone, BreakReasonUnspecified, func(lease *OpLock) bool {
-		return lease.HasRead() || lease.HasWrite()
-	}); err != nil {
-		return err
-	}
-	lm.breakDelegations(handleKey, excludeOwner, func(deleg *Delegation) bool {
-		return true
-	})
-
-	return nil
-}
-
-// CheckAndBreakCachingForRead breaks write leases (to Read) and write delegations.
-// Read delegations and read leases coexist with reads.
-func (lm *Manager) CheckAndBreakCachingForRead(handleKey string, excludeOwner *LockOwner) error {
-	if err := lm.breakOpLocks(handleKey, excludeOwner, LeaseStateRead, BreakReasonUnspecified, func(lease *OpLock) bool {
-		return lease.HasWrite()
-	}); err != nil {
-		return err
-	}
-	lm.breakDelegations(handleKey, excludeOwner, func(deleg *Delegation) bool {
-		return deleg.DelegType == DelegTypeWrite
-	})
-
-	return nil
-}
-
-// CheckAndBreakLeasesForSMBOpen breaks conflicting leases for an SMB CREATE.
-//
-// Per MS-SMB2 3.3.5.9 / MS-FSA 2.1.5.17.1: When a new SMB open arrives,
-// existing leases that hold Write caching must be broken. Unlike cross-protocol
-// breaks (CheckAndBreakCachingForWrite), the break strips only the Write bit,
-// preserving Read and Handle caching. This allows clients to continue read
-// and handle caching while flushing dirty data.
-//
-//   - RWH -> RH (strip W, keep Read + Handle)
-//   - RW  -> R  (strip W, keep Read)
-//   - R   -> not broken (no Write to strip)
-//   - RH  -> not broken (no Write to strip)
-func (lm *Manager) CheckAndBreakLeasesForSMBOpen(handleKey string, excludeOwner *LockOwner) error {
-	return lm.breakOpLocks(handleKey, excludeOwner, BreakToStripWrite, BreakReasonUnspecified, func(lease *OpLock) bool {
-		return lease.HasWrite()
-	})
-}
-
-// BreakLeasesForByteRangeLock breaks every other-key lease that holds Read
-// caching to None when an SMB byte-range lock is acquired.
-//
-// Per MS-SMB2 3.3.5.14 (Receiving an SMB2 LOCK Request) and Samba
-// `source3/smbd/smb2_oplock.c::contend_level2_oplocks_begin_default`
-// (lines 1391-1467) + `do_break_lease_to_none` (lines 1155-1206):
-// when a byte-range lock is granted on an open, every other lease holder
-// whose state has Read caching must be broken to None — Read caching
-// becomes invalid because the locking client may now write data the
-// remote cache can no longer observe.
-//
-// The locker's own lease (typically same lease key, possibly a same-key
-// secondary handle) is excluded via excludeOwner.ExcludeLeaseKey, mirroring
-// Samba's `smb2_lease_equal` no-self-break check.
-//
-// Leases without Read caching (None, or Write-only, which the protocol
-// disallows in practice) are skipped: there is no read cache to flush.
-// The break target is None — full revocation — not "strip W" or "strip H".
-func (lm *Manager) BreakLeasesForByteRangeLock(handleKey string, excludeOwner *LockOwner) error {
-	// An outstanding NFSv4 delegation is recalled for the same reason: the
-	// delegated client handles byte-range locks locally, so while it holds the
-	// delegation it can neither see this lock nor have its own locks seen. The
-	// delegation's whole-file row also blocks the lock in AddUnifiedLock, so
-	// without the recall the lock stays denied until the delegation's lease
-	// expires — even after the original holder released.
-	lm.breakDelegations(handleKey, excludeOwner, func(*Delegation) bool {
-		return true
-	})
-
-	return lm.breakOpLocks(handleKey, excludeOwner, LeaseStateNone, BreakReasonUnspecified, func(lease *OpLock) bool {
-		return lease.HasRead()
-	})
-}
-
-// BreakLeasesOnOpenConflict breaks leases held by other clients when an SMB
-// CREATE arrives. Per MS-SMB2 3.3.4.7 and Samba
-// `source3/smbd/open.c::delay_for_oplock_fn`. Per-lease target state is
-// computed via ComputeLeaseBreakTo(state, reason); a lease is broken only
-// when the computed target differs from its current state.
-func (lm *Manager) BreakLeasesOnOpenConflict(handleKey string, excludeOwner *LockOwner, reason BreakReason) error {
-	return lm.breakOpLocks(handleKey, excludeOwner, breakSentinelForReason(reason), reason, func(lease *OpLock) bool {
-		return ComputeLeaseBreakTo(lease.LeaseState, reason) != lease.LeaseState
-	})
-}
-
-// BreakReadLeasesForParentDir breaks Read leases on a parent directory when
-// a child file is modified (SET_INFO, WRITE, DELETE). Per MS-FSA 2.1.5.14:
-// changes to directory contents or child metadata invalidate Read caching,
-// so clients holding R or RW leases on the directory must be notified.
-//
-// The break goes to None (full revocation):
-//   - R  -> None
-//   - RW -> None
-func (lm *Manager) BreakReadLeasesForParentDir(handleKey string, excludeOwner *LockOwner) error {
-	return lm.breakOpLocks(handleKey, excludeOwner, LeaseStateNone, BreakReasonUnspecified, func(lease *OpLock) bool {
-		return lease.IsDirectory && lease.HasRead()
-	})
-}
-
-// CheckAndBreakCachingForDelete breaks all leases AND all delegations.
-func (lm *Manager) CheckAndBreakCachingForDelete(handleKey string, excludeOwner *LockOwner) error {
-	if err := lm.breakOpLocks(handleKey, excludeOwner, LeaseStateNone, BreakReasonUnspecified, func(lease *OpLock) bool {
-		return lease.LeaseState != LeaseStateNone
-	}); err != nil {
-		return err
-	}
-	lm.breakDelegations(handleKey, excludeOwner, func(deleg *Delegation) bool {
-		return true
-	})
-
-	return nil
-}
-
-// WaitForBreakCompletionExceptKey is WaitForBreakCompletion scoped to ignore
-// any breaking lease whose LeaseKey matches exceptKey. Used by the SMB CREATE
-// path on a same-key reopen: MS-SMB2 3.3.5.9.8 requires the opener to observe
-// Breaking=true on its own key (to emit SMB2_LEASE_FLAG_BREAK_IN_PROGRESS),
-// which forceCompleteBreaks would otherwise clear — but any other-key breaks
-// still need to drain before the CREATE proceeds (MS-SMB2 3.3.4.7). On
-// timeout, own-key is preserved and only other-key leases are force-completed.
-func (lm *Manager) WaitForBreakCompletionExceptKey(ctx context.Context, handleKey string, exceptKey [16]byte) error {
-	for {
-		lm.mu.Lock()
-		hasOther := false
-		for _, l := range lm.unifiedLocks[handleKey] {
-			if l.Lease != nil && l.Lease.Breaking && l.Lease.LeaseKey != exceptKey {
-				hasOther = true
-				break
-			}
-			if l.Delegation != nil && l.Delegation.Breaking {
-				hasOther = true
-				break
-			}
-		}
-		if !hasOther {
-			lm.mu.Unlock()
-			return nil
-		}
-
-		ch, ok := lm.breakWaitChans[handleKey]
-		if !ok {
-			ch = make(chan struct{})
-			lm.breakWaitChans[handleKey] = ch
-		}
-		lm.mu.Unlock()
-
-		select {
-		case <-ctx.Done():
-			lm.forceCompleteBreaksExceptKey(handleKey, exceptKey)
-			return ctx.Err()
-		case <-ch:
-			continue
-		}
-	}
-}
-
-// WaitForByteRangeLeaseBreak blocks until every breaking SMB lease on handleKey
-// resolves, IGNORING in-flight NFSv4 delegation breaks. It is the wait used by
-// the byte-range-lock acquisition path (NFSv4 LOCK / NLM), which runs while the
-// caller holds the NFSv4 StateManager mutex.
-//
-// It deliberately does NOT wait for delegation breaks, unlike
-// WaitForBreakCompletionExceptKey. A delegation break is resolved by
-// DELEGRETURN, whose handler must take the StateManager mutex; waiting for it
-// here while that same mutex is held would be a circular wait (only force-
-// broken after the bounded timeout, then re-stalled on the next LOCK, since
-// forceCompleteBreaksExceptKey only force-completes leases). Not waiting stays
-// correct: BreakLeasesForByteRangeLock has already dispatched the recall by the
-// time this runs, and the delegation's whole-file row keeps AddUnifiedLock
-// denying the acquire until the client returns it, so the lock is retried
-// rather than granted alongside an unrecalled delegation. On timeout the
-// breaking leases are force-downgraded to None (Samba
-// lease_timeout parity) and ctx.Err() is returned.
-func (lm *Manager) WaitForByteRangeLeaseBreak(ctx context.Context, handleKey string) error {
-	for {
-		lm.mu.Lock()
-		hasBreakingLease := false
-		for _, l := range lm.unifiedLocks[handleKey] {
-			if l.Lease != nil && l.Lease.Breaking {
-				hasBreakingLease = true
-				break
-			}
-		}
-		if !hasBreakingLease {
-			lm.mu.Unlock()
-			return nil
-		}
-
-		ch, ok := lm.breakWaitChans[handleKey]
-		if !ok {
-			ch = make(chan struct{})
-			lm.breakWaitChans[handleKey] = ch
-		}
-		lm.mu.Unlock()
-
-		select {
-		case <-ctx.Done():
-			// Force-break the holder's lease only when our own wait deadline
-			// expired (the holder genuinely failed to ACK in time). On a client
-			// cancellation no byte-range lock will be inserted, so we must not
-			// downgrade the SMB holder's lease as a side effect.
-			if errors.Is(ctx.Err(), context.DeadlineExceeded) {
-				lm.forceCompleteBreaksExceptKey(handleKey, [16]byte{})
-			}
-			return ctx.Err()
-		case <-ch:
-			continue
-		}
-	}
-}
-
-// byteRangeLockBreakWaitTimeout bounds how long a byte-range LOCK (NFSv4 or NLM)
-// waits for a conflicting SMB write-lease to finish breaking before it inserts
-// the lock. It mirrors the SMB lease-break ack-wait timeout: a holder that has
-// not ACKed within this window is treated as non-responsive, its lease is
-// force-downgraded to None (Samba lease_timeout parity), and the lock is granted.
-const byteRangeLockBreakWaitTimeout = 5 * time.Second
-
-// WaitForByteRangeLockBreak drains an in-flight lease break before a byte-range
-// LOCK inserts its lock. Callers must invoke it between
-// BreakLeasesForByteRangeLock and AddUnifiedLock.
-//
-// BreakLeasesForByteRangeLock is fire-and-forget: it marks the conflicting lease
-// Breaking and dispatches the break asynchronously, but the lease keeps its state
-// until the holder ACKs. The byte-lock conflict check gates on the lease's Write
-// bit, not on the Breaking flag, so inserting the lock before the break drains
-// returns a spurious DENIED (surfaced to the client as EIO) — the #1501
-// regression.
-//
-// Deadlock safety: it waits via WaitForByteRangeLeaseBreak, which is leases-only
-// (it never blocks on an NFSv4 delegation break — DELEGRETURN needs the
-// StateManager mutex this path holds) and releases the lock-manager mutex before
-// blocking, so it cannot deadlock against the break dispatch or the SMB
-// LEASE_BREAK_ACK path. The wait is bounded by byteRangeLockBreakWaitTimeout so a
-// stuck holder cannot stall the lock indefinitely; on expiry the breaking lease
-// is force-downgraded to None and the lock is granted. The empty exclude key
-// means "no exclusion": NFS/NLM owners hold no SMB lease of their own.
-//
-// It returns a non-nil error only when the originating ctx itself was cancelled
-// (e.g. the client disconnected) — the caller must then abort rather than insert
-// a lock nobody is waiting for. A derived deadline expiry returns nil: the lease
-// was force-downgraded above, so granting the lock is correct.
-func WaitForByteRangeLockBreak(ctx context.Context, lm LockManager, handleKey string) error {
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	waitCtx, cancel := context.WithTimeout(ctx, byteRangeLockBreakWaitTimeout)
-	defer cancel()
-	switch err := lm.WaitForByteRangeLeaseBreak(waitCtx, handleKey); {
-	case err == nil:
-		// Break drained (ACK or no breaking lease) — caller may insert the lock.
-		return nil
-	case ctx.Err() != nil:
-		// The originating request was cancelled/expired; abort the lock so we
-		// don't insert one nobody is waiting for.
-		return ctx.Err()
-	case errors.Is(err, context.DeadlineExceeded):
-		// Our derived deadline expired: the holder never ACKed, the lease was
-		// force-downgraded to None, and the lock may now be granted.
-		return nil
-	default:
-		// Unexpected error from the wait implementation — surface it.
-		return err
-	}
-}
-
-// AnyHolderHasLeaseBits reports whether any lease on handleKey (other than
-// exceptKey) currently has any bit in mask set. Non-blocking. Used by the SMB
-// CREATE post-break park decision per Samba `delay_for_oplock_fn`: a new opener
-// only needs to wait for the in-flight break ACK when the existing holder's
-// lease type intersects the delay_mask. Zero exceptKey means "no exclusion".
-func (lm *Manager) AnyHolderHasLeaseBits(handleKey string, exceptKey [16]byte, mask uint32) bool {
-	if mask == 0 {
-		return false
-	}
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-	hasExclusion := exceptKey != ([16]byte{})
-	for _, l := range lm.unifiedLocks[handleKey] {
-		if l.Lease == nil {
-			continue
-		}
-		if hasExclusion && l.Lease.LeaseKey == exceptKey {
-			continue
-		}
-		if l.Lease.LeaseState&mask != 0 {
-			return true
-		}
-	}
-	return false
-}
-
-// HasActiveLeaseRecord reports whether handleKey has any lease record (other
-// than one keyed on excludeKey) that is not a timeout tombstone. A holder
-// kept alive at LeaseState=None after ack-to-None still counts as active —
-// Samba's `disallow_write_lease` predicate (source3/smbd/open.c lines
-// 2397-2403) gates on `op_type != NO_OPLOCK`, not on lease state. Timeout
-// tombstones (BrokenViaTimeout=true) are excluded so a new opener after the
-// abandoned holder's timeout is not constrained by the dead record.
-func (lm *Manager) HasActiveLeaseRecord(handleKey string, excludeKey [16]byte) bool {
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-	for _, l := range lm.unifiedLocks[handleKey] {
-		if l.Lease == nil {
-			continue
-		}
-		if l.Lease.LeaseKey == excludeKey {
-			continue
-		}
-		if l.Lease.BrokenViaTimeout {
-			continue
-		}
-		return true
-	}
-	return false
-}
-
-// IsLeaseBrokenViaTimeout reports whether the lease identified by (handleKey,
-// leaseKey) is a timeout tombstone — its break was force-completed because the
-// holder never acknowledged it. The handle may still be open; the lease itself
-// is dead. The SMB CREATE W-strip uses this to tell a deadbeat same-client
-// lease (smb2.lease.timeout: keep WRITE off the new grant) from an active
-// same-client lease being upgraded (upgrade2/upgrade3: bypass the strip).
-//
-// Scoped to handleKey so a lease key bound on more than one file resolves to
-// the record on the file being granted, not an arbitrary same-key match.
-func (lm *Manager) IsLeaseBrokenViaTimeout(handleKey string, leaseKey [16]byte) bool {
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-	for _, l := range lm.unifiedLocks[handleKey] {
-		if l.Lease != nil && l.Lease.LeaseKey == leaseKey {
-			return l.Lease.BrokenViaTimeout
-		}
-	}
-	return false
-}
-
-// AnyHolderIsTraditionalOplock reports whether any record on handleKey is a
-// traditional oplock (IsTraditionalOplock=true). Used by the SMB CREATE path
-// to apply the narrower oplock stat-open mask when a traditional holder is
-// present (MS-SMB2 §3.3.5.9 / Samba `is_oplock_stat_open`).
-func (lm *Manager) AnyHolderIsTraditionalOplock(handleKey string) bool {
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-	for _, l := range lm.unifiedLocks[handleKey] {
-		if l.Lease != nil && l.Lease.IsTraditionalOplock {
-			return true
-		}
-	}
-	return false
-}
-
-// OnlyTimeoutTombstoneRecords reports whether handleKey has at least one
-// lease record AND every present lease record is a timeout tombstone
-// (BrokenViaTimeout=true). Returns false when no records exist at all, or
-// when at least one record is not a timeout tombstone.
-//
-// Used by the CREATE-grant LEVEL_II coercion to distinguish "holder timed
-// out and the server moved on" (only timeout tombstones present → don't
-// constrain the new grant by the abandoned holder) from "holder normally
-// acked or is still active" (at least one live record → defer to
-// bestGrantableState or fall back to non-stat-open coercion).
-//
-// Covers the contrast between smbtorture batch22b (timeout → tree2 BATCH
-// expected) and exclusive9 SUPERSEDE iteration (ack → tree2 LEVEL_II
-// expected) which both leave LeaseState=None records but originate from
-// different paths.
-func (lm *Manager) OnlyTimeoutTombstoneRecords(handleKey string) bool {
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-	any := false
-	for _, l := range lm.unifiedLocks[handleKey] {
-		if l.Lease == nil {
-			continue
-		}
-		any = true
-		if !l.Lease.BrokenViaTimeout {
-			return false
-		}
-	}
-	return any
-}
-
-// HasOtherBreakingLeases reports whether any lease (other than exceptKey) or
-// any delegation on handleKey is currently Breaking. Non-blocking. Used by the
-// SMB CREATE async-park path to decide whether to emit STATUS_PENDING and
-// resume the CREATE from a goroutine. A zero exceptKey means "no exclusion" —
-// any Breaking lease matches.
-func (lm *Manager) HasOtherBreakingLeases(handleKey string, exceptKey [16]byte) bool {
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-	hasExclusion := exceptKey != ([16]byte{})
-	for _, l := range lm.unifiedLocks[handleKey] {
-		if l.Lease != nil && l.Lease.Breaking {
-			if !hasExclusion || l.Lease.LeaseKey != exceptKey {
-				return true
-			}
-		}
-		if l.Delegation != nil && l.Delegation.Breaking {
-			return true
-		}
-	}
-	return false
-}
-
-// WaitForBreakCompletion blocks until all breaking locks on a file resolve
-// or the context is cancelled. Multiple goroutines may wait concurrently;
-// signalBreakWait uses close() to broadcast to all waiters.
-//
-// On timeout (context cancellation), all leases still in Breaking state are
-// automatically downgraded to their BreakToState, as if the client had
-// acknowledged. Per MS-SMB2 3.3.5.22.2: if the client fails to acknowledge
-// within the timeout, the server completes the break.
-func (lm *Manager) WaitForBreakCompletion(ctx context.Context, handleKey string) error {
-	for {
-		lm.mu.Lock()
-		hasBreaking := false
-		for _, lock := range lm.unifiedLocks[handleKey] {
-			if lock.Lease != nil && lock.Lease.Breaking {
-				hasBreaking = true
-				break
-			}
-			if lock.Delegation != nil && lock.Delegation.Breaking {
-				hasBreaking = true
-				break
-			}
-		}
-
-		if !hasBreaking {
-			lm.mu.Unlock()
-			return nil
-		}
-
-		// Get or create the wait channel while still holding the lock,
-		// so no signal from signalBreakWait can be missed.
-		ch, ok := lm.breakWaitChans[handleKey]
-		if !ok {
-			ch = make(chan struct{})
-			lm.breakWaitChans[handleKey] = ch
-		}
-		lm.mu.Unlock()
-
-		select {
-		case <-ctx.Done():
-			// Timeout: auto-downgrade all breaking leases to their break-to state.
-			lm.forceCompleteBreaks(handleKey)
-			return ctx.Err()
-		case <-ch:
-			continue
-		}
-	}
-}
-
-// WaitForShareConflictClear blocks until one of three conditions, whichever
-// comes first, then returns. It re-evaluates conflictPresent on every break-wait
-// signal AND on a short poll:
-//
-//   - conflictPresent() reports false → the holder CLOSEd (its open-table entry
-//     is gone), the conflict cleared → returns nil; the caller's CREATE proceeds.
-//   - the holder's break drained but the conflict is still present → the holder
-//     ACKed the break while keeping its open (e.g. RWH→RW) → returns nil; the
-//     caller's final recheck then yields SHARING_VIOLATION. This early exit is
-//     deterministic, so ack-sane does not stall to the deadline.
-//   - ctx is cancelled (genuine never-released conflict) → returns ctx.Err();
-//     the caller's recheck yields SHARING_VIOLATION against the still-live holder.
-//
-// A nil return therefore means "stop waiting and recheck", NOT "conflict
-// cleared" — the caller MUST re-run the share-mode check to decide the outcome.
-//
-// Crucially, unlike WaitForBreakCompletion, NONE of these paths force-complete
-// (tombstone) the holder's breaking lease: the deferred-open contract
-// (MS-SMB2 §3.3.5.9, Samba defer_open→retry_open) lets the holder ack on its own
-// schedule, and forcing the lease to None here would make a later ACK fail
-// STATUS_UNSUCCESSFUL (smbtorture dhv2-pending1n-vs-violation-lease-ack-sane).
-func (lm *Manager) WaitForShareConflictClear(ctx context.Context, handleKey string, conflictPresent func() bool) error {
-	// Poll interval. A holder CLOSE that resolves the conflict does NOT signal
-	// the breakWaitChans channel for file leases (that signal is scoped to
-	// directory leases so a file-lease holder closing without ACKing cannot
-	// prematurely wake a regular parked CREATE — smbtorture
-	// smb2.kernel-oplocks.kernel_oplocks7). The deferred-open resume instead
-	// re-evaluates the live share-mode predicate on a short poll, plus on every
-	// break-wait signal, so it wakes promptly on either a CLOSE or an ACK
-	// without disturbing the regular break-completion wait path.
-	const pollInterval = 100 * time.Millisecond
-	ticker := time.NewTicker(pollInterval)
-	defer ticker.Stop()
-
-	for {
-		if !conflictPresent() {
-			return nil
-		}
-
-		lm.mu.Lock()
-		// Early deterministic exit: the conflict is still present but no break
-		// is in flight any more. The holder resolved its break by ACKing
-		// (keeping its open, e.g. RWH→RW) rather than CLOSEing, so the share
-		// conflict will never clear on its own. Return now so the caller's
-		// final recheck yields SHARING_VIOLATION promptly instead of stalling
-		// until the deadline (smbtorture dhv2-pending1n-vs-violation-lease-ack-sane).
-		if !lm.hasBreakingLeaseLocked(handleKey) {
-			lm.mu.Unlock()
-			return nil
-		}
-		ch, ok := lm.breakWaitChans[handleKey]
-		if !ok {
-			ch = make(chan struct{})
-			lm.breakWaitChans[handleKey] = ch
-		}
-		lm.mu.Unlock()
-
-		// Re-check after subscribing so a signal racing between the predicate
-		// evaluation above and channel registration is not missed.
-		if !conflictPresent() {
-			return nil
-		}
-
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-ch:
-			continue
-		case <-ticker.C:
-			continue
-		}
-	}
-}
-
-// hasBreakingLeaseLocked reports whether any lease or delegation on handleKey is
-// currently in the Breaking state. Caller holds lm.mu.
-func (lm *Manager) hasBreakingLeaseLocked(handleKey string) bool {
-	for _, l := range lm.unifiedLocks[handleKey] {
-		if l.Lease != nil && l.Lease.Breaking {
-			return true
-		}
-		if l.Delegation != nil && l.Delegation.Breaking {
-			return true
-		}
-	}
-	return false
-}
-
-// forceCompleteBreaks force-revokes all breaking leases on a file to None when
-// the break wait times out. Records are kept alive at LeaseState=None
-// (handle-bound lifetime) so a later unsolicited ack surfaces as
-// ErrLeaseAckNotBreaking → STATUS_UNSUCCESSFUL.
-func (lm *Manager) forceCompleteBreaks(handleKey string) {
-	lm.forceCompleteBreaksExceptKey(handleKey, [16]byte{})
-}
-
-// forceCompleteBreaksExceptKey is forceCompleteBreaks that leaves any lease
-// keyed on exceptKey untouched. Zero exceptKey means "no exclusion" (same
-// semantics as forceCompleteBreaks).
-//
-// Forces breaking leases to None, mirroring Samba's lease_timeout_handler
-// (source3/smbd/smb2_oplock.c) which calls
-// `downgrade_lease(..., SMB2_LEASE_NONE)` regardless of the in-flight or
-// cumulative break target. A non-acking client must not be allowed to retain
-// any lease bits past the timeout — otherwise a later opener observes stale
-// state (smb2.lease.timeout: probe of original lease key returns RH instead
-// of the spec-mandated empty state) and stale R/H rights would generate
-// spurious break notifications on subsequent IO.
-func (lm *Manager) forceCompleteBreaksExceptKey(handleKey string, exceptKey [16]byte) {
-	lm.mu.Lock()
-	defer lm.mu.Unlock()
-
-	modified := false
-	for _, l := range lm.unifiedLocks[handleKey] {
-		if l.Lease == nil || !l.Lease.Breaking || l.Lease.LeaseKey == exceptKey {
-			continue
-		}
-		modified = true
-		// Force-revoke to None and keep the record alive at LeaseState=None
-		// (handle-bound lifetime) so a later unsolicited or duplicate ack
-		// surfaces as ErrLeaseAckNotBreaking → STATUS_UNSUCCESSFUL. Same
-		// rationale as applyBreakStageLocked.
-		//
-		// Do NOT advance Epoch: this is the timeout/internal completion
-		// path for an already-dispatched break. Per MS-SMB2 §3.3.4.7 the
-		// epoch advances only when a break notification is dispatched (and
-		// was already advanced when the in-flight break started). Bumping
-		// it here would invalidate any straggling client ack still echoing
-		// the original epoch.
-		l.Lease.LeaseState = LeaseStateNone
-		l.Lease.BreakingToRequired = LeaseStateNone
-		l.Lease.Breaking = false
-		l.Lease.BreakToState = 0
-		l.Lease.BreakStarted = time.Time{}
-		l.Lease.BrokenViaTimeout = true
-		l.Type = lockTypeForLeaseState(l.Lease.LeaseState)
-
-		lm.persistUnifiedLockLocked(l)
-		logger.Debug("forceCompleteBreaks: auto-downgraded lease",
-			"handleKey", handleKey,
-			"leaseKey", fmt.Sprintf("%x", l.Lease.LeaseKey),
-			"newState", LeaseStateToString(l.Lease.LeaseState))
-	}
-
-	if modified {
-		lm.signalBreakWaitLocked(handleKey)
-	}
-}
-
-// signalBreakWait broadcasts to all waiters by closing the wait channel and
-// removing it from the map. The next WaitForBreakCompletion call will create
-// a fresh channel if needed. Acquires lm.mu internally.
-func (lm *Manager) signalBreakWait(handleKey string) {
-	lm.mu.Lock()
-	lm.signalBreakWaitLocked(handleKey)
-	lm.mu.Unlock()
-}
-
-// SignalParkedCreates is the LockManager-interface entry point for
-// signalBreakWait, exposed so the SMB CLOSE path can wake a parked CREATE
-// waiter after the open-file table entry has been removed. See interface doc.
-func (lm *Manager) SignalParkedCreates(handleKey string) {
-	lm.signalBreakWait(handleKey)
-}
-
-// signalBreakWaitLocked is the lock-held variant of signalBreakWait.
-// Caller must hold lm.mu.
-func (lm *Manager) signalBreakWaitLocked(handleKey string) {
-	if ch, ok := lm.breakWaitChans[handleKey]; ok {
-		close(ch)
-		delete(lm.breakWaitChans, handleKey)
-	}
-}
-
-// breakOpLocks marks matching oplocks as breaking and dispatches break
-// notifications. Releases mutex before dispatching to avoid deadlock.
-//
-// breakToState is the target state for the break. Pass BreakToStripWrite
-// to compute the per-lease break-to state by stripping the Write bit from
-// each lease's current state (preserving Read and Handle).
-//
-// Concurrent-break behavior: when a lease is already Breaking, the new
-// target is AND-merged into BreakingToRequired (cumulative final target)
-// without dispatching a new notification or advancing the epoch. This
-// mirrors Samba `process_oplock_break_message` lines 956-965; the next
-// progressive stage is dispatched from acknowledgeLeaseBreakImpl after the
-// in-flight ACK arrives.
-func (lm *Manager) breakOpLocks(
-	handleKey string,
-	excludeOwner *LockOwner,
-	breakToState uint32,
-	reason BreakReason,
-	shouldBreak func(lease *OpLock) bool,
-) error {
-	lm.mu.Lock()
-	locks := lm.unifiedLocks[handleKey]
-
-	type breakEntry struct {
-		lock         *UnifiedLock
-		breakToState uint32
-	}
-	var toBreak []breakEntry
-	// Dedup wire break notifications by lease key. A handleKey can hold more
-	// than one *UnifiedLock with the same lease key — e.g. an orphaned
-	// ack-to-None record left by an unclean disconnect coexisting with a live
-	// regrant under the smbtorture-reused DLEASE constant. Samba dispatches
-	// exactly one LEASE_BREAK per distinct holder/lease key per change
-	// (source3/smbd/smb2_oplock.c contend_dirleases); dispatching once per
-	// record sends duplicate notifications that all route to the same live
-	// session via the shared ClientGUID, producing the intermittent
-	// smb2.dirlease.unlink_*_and_close "lease_break_info.count got 0x2" flake.
-	dispatchedKeys := make(map[[16]byte]struct{}, len(locks))
-	// canonicalByKey remembers the record whose break stage drove the single
-	// wire notification for each lease key. A sibling sharing that key (an
-	// orphaned ack-to-None record left by an unclean disconnect coexisting
-	// with the live regrant) must NOT send a second notification, but its
-	// internal break stage MUST be mirrored from the canonical record so a
-	// later scan can't treat the stale sibling as an active non-breaking
-	// lease and dispatch a fresh spurious break (per MS-SMB2 §3.3.5.9: opens
-	// sharing a lease key share one logical lease).
-	canonicalByKey := make(map[[16]byte]*OpLock, len(locks))
-	for _, lock := range locks {
-		if lock.Lease == nil {
-			continue
-		}
-		if excludeOwner != nil {
-			if lock.Owner.OwnerID == excludeOwner.OwnerID ||
-				(excludeOwner.ClientID != "" && lock.Owner.ClientID == excludeOwner.ClientID) {
-				continue
-			}
-			// Per MS-SMB2 3.3.5.9: opens with the same lease key must not
-			// break each other's leases ("If Open.Lease.LeaseKey == the new
-			// open's LeaseKey, no break is required").
-			if excludeOwner.ExcludeLeaseKey != ([16]byte{}) &&
-				lock.Lease.LeaseKey == excludeOwner.ExcludeLeaseKey {
-				continue
-			}
-			// Per MS-SMB2 §3.3.4.20 / Samba dirlease_should_break: when a
-			// child CREATE or SET_INFO carries an RqLs whose ParentLeaseKey
-			// matches the parent dir lease's key, suppress that dir-lease
-			// break. Scoped to dir leases to avoid suppressing an unrelated
-			// file-lease break that happens to share the key value (#470 C2).
-			if excludeOwner.HasExcludeParentDirLeaseKey &&
-				lock.Lease.IsDirectory &&
-				lock.Lease.LeaseKey == excludeOwner.ExcludeParentDirLeaseKey {
-				continue
-			}
-		}
-		if !shouldBreak(lock.Lease) {
-			continue
-		}
-
-		// One notification per distinct lease key (see dispatchedKeys above):
-		// a duplicate record under this handleKey must not produce a second
-		// wire break for the same lease. Marked once the key is accepted for
-		// a fresh dispatch below so the live record drives the notification and
-		// any sibling sharing its key is skipped.
-		if _, already := dispatchedKeys[lock.Lease.LeaseKey]; already {
-			// Mirror the canonical record's break stage onto this sibling
-			// (no wire notification, no second epoch advance) so its state
-			// stays consistent with the one logical lease.
-			if canonical := canonicalByKey[lock.Lease.LeaseKey]; canonical != nil {
-				lm.mirrorBreakStageLocked(lock, canonical, true)
-			}
-			continue
-		}
-
-		targetState := computeFreshTarget(lock.Lease.LeaseState, breakToState)
-
-		// Per Samba `delay_for_oplock_fn` (source3/smbd/open.c lines 2439-2444):
-		// traditional oplocks only support breaking to R or NONE — any Handle
-		// or Write residue in the strip-W/strip-H target must be cleared so the
-		// holder lands at R (LEVEL_II) or 0 (NONE). Lease holders retain the
-		// fine-grained break-to bits; this mask only applies to traditional
-		// oplocks tagged at grant time. Required by smbtorture
-		// smb2.oplock.batch9a (BATCH attrs-only holder must break to R so the
-		// subsequent normal-open BATCH request can be granted LEVEL_II via
-		// bestGrantableState).
-		if lock.Lease.IsTraditionalOplock {
-			targetState &^= LeaseStateHandle | LeaseStateWrite
-		}
-
-		if lock.Lease.Breaking {
-			// Concurrent break: AND-merge the new opener's target into the
-			// cumulative final target. No notification, no epoch bump
-			// (Samba intentionally skips the bump per its inline comment).
-			// The next progressive stage will be dispatched on ACK.
-			lock.Lease.BreakingToRequired &= targetState
-			// Upgrade the recorded reason if this opener classifies the break
-			// (e.g. an open-conflict lease downgrade merging into an in-flight
-			// fire-and-forget break); never clobber a real reason with Unspecified.
-			if reason != BreakReasonUnspecified {
-				lock.Lease.BreakReason = reason
-			}
-			lm.persistUnifiedLockLocked(lock)
-			// This key's break is already in flight; record it so a sibling
-			// record sharing the key is not freshly dispatched below and so
-			// later siblings mirror this record's break stage.
-			dispatchedKeys[lock.Lease.LeaseKey] = struct{}{}
-			canonicalByKey[lock.Lease.LeaseKey] = lock.Lease
-			continue
-		}
-
-		// Fresh dispatch: BreakingToRequired starts at this opener's target.
-		// Subsequent concurrent breaks may tighten it via the AND-merge above.
-		// Advance the epoch here so the dispatched notification's NewEpoch is
-		// pre-bumped (per MS-SMB2 2.2.23.2). Post-ACK progressive stages do
-		// NOT advance — the multi-stage progression is one logical break.
-		lock.Lease.BreakingToRequired = targetState
-		lock.Lease.BreakReason = reason
-		advanceEpoch(lock.Lease)
-		snapshot := lm.applyBreakStageLocked(lock, targetState)
-		// Persist the in-flight Breaking state so a crash/restart preserves
-		// the break-in-progress and parked CREATEs aren't stranded waiting
-		// for a notification that was already sent over the wire.
-		// applyBreakStageLocked only persists the fire-and-forget downgrade
-		// path; the ack-required path (which is the common case) is
-		// persisted here.
-		lm.persistUnifiedLockLocked(lock)
-		dispatchedKeys[lock.Lease.LeaseKey] = struct{}{}
-		canonicalByKey[lock.Lease.LeaseKey] = lock.Lease
-		toBreak = append(toBreak, breakEntry{lock: snapshot, breakToState: targetState})
-	}
-	lm.mu.Unlock()
-
-	for _, entry := range toBreak {
-		lm.dispatchOpLockBreak(handleKey, entry.lock, entry.breakToState)
-	}
-
-	return nil
-}
-
-// computeFreshTarget resolves a breakOpLocks sentinel against the current
-// lease state, returning the actual per-lease target. Direct state values
-// pass through unchanged.
-func computeFreshTarget(currentState, sentinel uint32) uint32 {
-	switch sentinel {
-	case BreakToStripWrite:
-		// Per MS-SMB2 3.3.5.9: RWH -> RH, RW -> R.
-		return currentState &^ LeaseStateWrite
-	case BreakToStripHandle:
-		// Per MS-SMB2 3.3.5.9 Step 10: RWH -> RW, RH -> R.
-		return currentState &^ LeaseStateHandle
-	}
-	return sentinel
-}
-
-// applyBreakStageLocked performs a single break stage on lock targeting
-// `target`. Caller must hold lm.mu, must have already set
-// lock.Lease.BreakingToRequired appropriately, and is responsible for
-// dispatching the returned snapshot via dispatchOpLockBreak after releasing
-// lm.mu.
-//
-// Per MS-SMB2 3.3.4.7, a break is ack-required iff the current state is NOT
-// pure Read. Without ACK_REQUIRED the client never responds, so leaving
-// Breaking=true would block same-key reopens — instead we resolve inline.
-//
-// For target=None on the inline (fire-and-forget) path we keep the record
-// alive at LeaseState=None (handle-bound lifetime) so a later unsolicited
-// ack from the client surfaces as ErrLeaseAckNotBreaking →
-// STATUS_UNSUCCESSFUL per MS-SMB2 3.3.5.22.2 (smbtorture breaking5). The
-// record is removed when the holding handle CLOSEs (ReleaseLeaseForHandle).
-func (lm *Manager) applyBreakStageLocked(lock *UnifiedLock, target uint32) *UnifiedLock {
-	// Snapshot while LeaseState still holds the pre-break value for
-	// CurrentLeaseState in the notification. Caller is responsible for
-	// advancing the epoch on fresh dispatch (per MS-SMB2 2.2.23.2). Progressive
-	// next-stage dispatch from a post-ACK re-eval does NOT advance epoch — the
-	// multi-stage break is one continuous progression and Samba's
-	// `downgrade_lease` (source3/smbd/smb2_oplock.c line 607) reuses the
-	// existing epoch unchanged for each intermediate stage.
-	snapshot := lock.Clone()
-
-	ackRequired := lock.Lease.LeaseState != LeaseStateRead
-	if ackRequired {
-		lock.Lease.Breaking = true
-		lock.Lease.BreakToState = target
-		lock.Lease.BreakStarted = time.Now()
-		return snapshot
-	}
-
-	// Fire-and-forget downgrade: client won't ACK (current state is pure R).
-	lock.Lease.Breaking = false
-	lock.Lease.BreakToState = 0
-	lock.Lease.BreakStarted = time.Time{}
-	lock.Lease.LeaseState = target
-	lock.Lease.BreakingToRequired = target
-	lock.Type = lockTypeForLeaseState(target)
-	lm.persistUnifiedLockLocked(lock)
-	return snapshot
-}
-
-// mirrorBreakStageLocked copies the break-stage fields of a canonical lease
-// record onto a sibling that shares its lease key, WITHOUT sending a wire
-// notification or independently advancing the sibling's epoch. Per MS-SMB2
-// §3.3.5.9 opens sharing a lease key share one logical lease, so the sibling
-// must end in the same break stage (Breaking / BreakToState / BreakingToRequired
-// / BreakStarted / Epoch / LeaseState) as the canonical record. Without this,
-// the skipped sibling keeps stale pre-break values and a later scan could treat
-// it as an active non-breaking lease and dispatch a fresh spurious break.
-//
-// The canonical record's epoch was advanced exactly once on its fresh dispatch
-// (or intentionally not advanced on the concurrent-break merge path); copying it
-// keeps the sibling in lockstep rather than diverging via a second advance.
-//
-// When persist is true the sibling is persisted (the caller persists the
-// canonical too); callers whose canonical path does not persist pass false so
-// the sibling stays in lockstep with the canonical's durability. Caller must
-// hold lm.mu.
-func (lm *Manager) mirrorBreakStageLocked(sibling *UnifiedLock, canonical *OpLock, persist bool) {
-	sibling.Lease.Breaking = canonical.Breaking
-	sibling.Lease.BreakToState = canonical.BreakToState
-	sibling.Lease.BreakingToRequired = canonical.BreakingToRequired
-	sibling.Lease.BreakStarted = canonical.BreakStarted
-	sibling.Lease.Epoch = canonical.Epoch
-	sibling.Lease.LeaseState = canonical.LeaseState
-	sibling.Type = lockTypeForLeaseState(canonical.LeaseState)
-	if persist {
-		lm.persistUnifiedLockLocked(sibling)
-	}
-}
-
-// ============================================================================
-// Delegation CRUD Operations
-// ============================================================================
-
-// GrantDelegation grants a delegation on a file.
-// Returns error if a conflicting lease or byte-range lock exists, or the file
-// was recently broken.
-func (lm *Manager) GrantDelegation(handleKey string, delegation *Delegation) error {
-	lm.mu.Lock()
-	defer lm.mu.Unlock()
-
-	// Check anti-storm cache inside the lock to be atomic with lease conflict check.
-	if lm.recentlyBroken != nil && lm.recentlyBroken.IsRecentlyBroken(handleKey) {
-		return fmt.Errorf("delegation denied: file recently had caching broken")
-	}
-
-	locks := lm.unifiedLocks[handleKey]
-
-	// Check lease conflicts. Delegation-vs-delegation conflicts (e.g., at most
-	// one write delegation per file) are enforced by the protocol layer (NFS
-	// state manager, SMB handler) before calling GrantDelegation.
-	for _, lock := range locks {
-		if lock.Lease != nil && DelegationConflictsWithLease(delegation, lock.Lease) {
-			return fmt.Errorf("delegation conflicts with existing lease (state=%s)",
-				LeaseStateToString(lock.Lease.LeaseState))
-		}
-	}
-
-	newLock := &UnifiedLock{
-		ID: delegation.DelegationID,
-		Owner: LockOwner{
-			OwnerID:   DelegationOwnerID(delegation.ClientID, delegation.DelegationID),
-			ClientID:  delegation.ClientID,
-			ShareName: delegation.ShareName,
-		},
-		FileHandle: FileHandle(handleKey),
-		Offset:     0,
-		Length:     0, // Whole file
-		Type:       delegationToLockType(delegation.DelegType),
-		AcquiredAt: time.Now(),
-		Delegation: delegation,
-	}
-
-	// A byte-range lock already held on this file denies the delegation. A
-	// delegated client is entitled to satisfy its own byte-range locks locally,
-	// without ever asking the server, so handing out a delegation over an
-	// existing lock makes that lock invisible to the delegated client — the
-	// cross-protocol conflict that AddUnifiedLock enforces in the other
-	// direction would simply never be evaluated.
-	//
-	// Both lock maps are consulted: NLM/NFSv4 locks live in unifiedLocks and
-	// SMB locks in locks. The delegation is weighed as the whole-file byte-range
-	// lock it stands in for, so a write delegation (exclusive) is denied by any
-	// foreign lock and a read delegation (shared) only by a foreign exclusive
-	// one — the same rule byte-range locks apply to each other, including its
-	// exemption for the holder's own client.
-	for _, lock := range locks {
-		if lock.Lease != nil || lock.Delegation != nil {
-			continue
-		}
-		if newLock.ConflictsWith(lock) {
-			return fmt.Errorf("delegation conflicts with existing byte-range lock (owner=%s)",
-				lock.Owner.OwnerID)
-		}
-	}
-	// The SMB map needs a plain byte-range stand-in: fileLockConflictsWithUnified
-	// excludes delegation rows by design. SMB locks never share a client with an
-	// NFS delegation, so no same-client exemption applies here.
-	asByteRange := *newLock
-	asByteRange.Delegation = nil
-	smbLocks := lm.locks[handleKey]
-	for i := range smbLocks {
-		if fileLockConflictsWithUnified(&smbLocks[i], &asByteRange) {
-			return fmt.Errorf("delegation conflicts with existing SMB byte-range lock (owner=%s)",
-				lockOwnerID(&smbLocks[i]))
-		}
-	}
-
-	lm.unifiedLocks[handleKey] = append(locks, newLock)
-	lm.indexAddLockLocked(handleKey, newLock)
-	return nil
-}
-
-// DelegationOwnerID returns the OwnerID that GrantDelegation assigns to a
-// delegation. This is useful for constructing an excludeOwner that matches
-// the delegation's LockOwner.
-func DelegationOwnerID(clientID, delegationID string) string {
-	return fmt.Sprintf("deleg:%s:%s", clientID, delegationID)
-}
-
-// delegationToLockType converts a DelegationType to a LockType.
-func delegationToLockType(dt DelegationType) LockType {
-	if dt == DelegTypeWrite {
-		return LockTypeExclusive
-	}
-	return LockTypeShared
-}
-
-// RevokeDelegation force-revokes a delegation, removing it from the lock map.
-func (lm *Manager) RevokeDelegation(handleKey string, delegationID string) error {
-	found := lm.removeMatchingLocksAndSignal(handleKey, false, func(l *UnifiedLock) bool {
-		return l.Delegation != nil && l.Delegation.DelegationID == delegationID
-	})
-	if !found {
-		return fmt.Errorf("delegation %s not found on handle %s", delegationID, handleKey)
-	}
-	return nil
-}
-
-// revokeTimedOutLease removes all in-memory lease records matching leaseKey
-// from handleKey, deletes their persisted records (best-effort), and signals
-// WaitForBreakCompletion waiters. Called by OpLockBreakScanner after the
-// persistent record has already been deleted by the scanner's DeleteLock call;
-// this method handles the in-memory half.
-//
-// This is the lease analogue of RevokeDelegation: same mutex discipline,
-// same signalBreakWait call so parked CREATEs unblock immediately rather
-// than waiting for their context deadline.
-func (lm *Manager) revokeTimedOutLease(handleKey string, leaseKey [16]byte) {
-	// Best-effort delete of each removed record's persisted copy: the scanner
-	// already deleted the canonical PersistedLock by ID; this handles any
-	// sibling records sharing the same LeaseKey.
-	lm.removeMatchingLocksAndSignal(handleKey, true, func(l *UnifiedLock) bool {
-		return l.Lease != nil && l.Lease.LeaseKey == leaseKey
-	})
-}
-
-// removeMatchingLocksAndSignal removes every UnifiedLock on handleKey for which
-// match returns true, then wakes any WaitForBreakCompletion / parked-CREATE
-// waiters. When deletePersisted is true, each removed lock's persisted record is
-// deleted (best-effort) under lm.mu. Returns true if at least one lock matched.
-//
-// Shared by RevokeDelegation and revokeTimedOutLease: same mutex discipline,
-// same post-unlock signalBreakWait so parked CREATEs unblock immediately rather
-// than waiting for their context deadline.
-func (lm *Manager) removeMatchingLocksAndSignal(handleKey string, deletePersisted bool, match func(*UnifiedLock) bool) bool {
-	lm.mu.Lock()
-	old := lm.unifiedLocks[handleKey]
-	var remaining []*UnifiedLock
-	found := false
-	for _, l := range old {
-		if match(l) {
-			found = true
-			if deletePersisted {
-				lm.deleteUnifiedLockLocked(l)
-			}
-			continue
-		}
-		remaining = append(remaining, l)
-	}
-	if !found {
-		lm.mu.Unlock()
-		return false
-	}
-	if len(remaining) == 0 {
-		delete(lm.unifiedLocks, handleKey)
-	} else {
-		lm.unifiedLocks[handleKey] = remaining
-	}
-	lm.reindexHandleLocked(handleKey, old)
-	lm.mu.Unlock()
-
-	lm.signalBreakWait(handleKey)
-	return true
-}
-
-// ReturnDelegation handles a client returning a delegation. Idempotent:
-// returns nil even if the delegation was not found.
-func (lm *Manager) ReturnDelegation(handleKey string, delegationID string) error {
-	lm.mu.Lock()
-
-	locks := lm.unifiedLocks[handleKey]
-	var remaining []*UnifiedLock
-	for _, l := range locks {
-		if l.Delegation != nil && l.Delegation.DelegationID == delegationID {
-			continue
-		}
-		remaining = append(remaining, l)
-	}
-
-	if len(remaining) == 0 {
-		delete(lm.unifiedLocks, handleKey)
-	} else {
-		lm.unifiedLocks[handleKey] = remaining
-	}
-	lm.reindexHandleLocked(handleKey, locks)
-	lm.mu.Unlock()
-
-	lm.signalBreakWait(handleKey)
-	return nil
-}
-
-// GetDelegation retrieves a specific delegation by ID.
-// Returns nil if not found.
-func (lm *Manager) GetDelegation(handleKey string, delegationID string) *Delegation {
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-
-	for _, lock := range lm.unifiedLocks[handleKey] {
-		if lock.Delegation != nil && lock.Delegation.DelegationID == delegationID {
-			return lock.Delegation.Clone()
-		}
-	}
-	return nil
-}
-
-// ListDelegations returns all delegations on a file.
-func (lm *Manager) ListDelegations(handleKey string) []*Delegation {
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-
-	var result []*Delegation
-	for _, lock := range lm.unifiedLocks[handleKey] {
-		if lock.Delegation != nil {
-			result = append(result, lock.Delegation.Clone())
-		}
-	}
-	return result
-}
-
-// ExpiredDelegation holds info about a delegation whose recall has timed out.
-type ExpiredDelegation struct {
-	HandleKey    string
-	DelegationID string
-}
-
-// CollectExpiredDelegationRecalls returns delegations that are in the breaking
-// state and have exceeded the given timeout. This allows external scanners to
-// query for expired recalls without accessing internal fields.
-func (lm *Manager) CollectExpiredDelegationRecalls(now time.Time, timeout time.Duration) []ExpiredDelegation {
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-
-	var expired []ExpiredDelegation
-	for handleKey, locks := range lm.unifiedLocks {
-		for _, lock := range locks {
-			if lock.Delegation == nil || !lock.Delegation.Breaking {
-				continue
-			}
-			if now.After(lock.Delegation.BreakStarted.Add(timeout)) {
-				expired = append(expired, ExpiredDelegation{
-					HandleKey:    handleKey,
-					DelegationID: lock.Delegation.DelegationID,
-				})
-			}
-		}
-	}
-	return expired
-}
-
-// breakDelegations collects delegations matching the predicate and dispatches
-// recall notifications. Releases mutex before dispatching to avoid deadlock.
-//
-// excludeOwner skips delegations whose Owner.OwnerID matches. Delegation
-// OwnerIDs use the format "deleg:{clientID}:{delegationID}". Callers that
-// want to exclude by client identity should match on Owner.ClientID instead,
-// or construct the OwnerID in the same format.
-func (lm *Manager) breakDelegations(
-	handleKey string,
-	excludeOwner *LockOwner,
-	shouldBreak func(deleg *Delegation) bool,
-) {
-	lm.mu.Lock()
-	locks := lm.unifiedLocks[handleKey]
-
-	var toBreak []*UnifiedLock
-	for _, lock := range locks {
-		if lock.Delegation == nil {
-			continue
-		}
-		if excludeOwner != nil &&
-			(lock.Owner.OwnerID == excludeOwner.OwnerID ||
-				(excludeOwner.ClientID != "" && lock.Owner.ClientID == excludeOwner.ClientID)) {
-			continue
-		}
-		if lock.Delegation.Breaking {
-			continue
-		}
-		if shouldBreak(lock.Delegation) {
-			lock.Delegation.Breaking = true
-			lock.Delegation.BreakStarted = time.Now()
-			// Clone before dispatch to prevent race with concurrent ack/release.
-			toBreak = append(toBreak, lock.Clone())
-		}
-	}
-	lm.mu.Unlock()
-
-	if len(toBreak) > 0 && lm.recentlyBroken != nil {
-		lm.recentlyBroken.Mark(handleKey)
-	}
-
-	for _, lock := range toBreak {
-		lm.dispatchDelegationRecall(handleKey, lock)
-	}
-}
-
-// dispatchDelegationRecall notifies all registered break callbacks about a delegation recall.
-func (lm *Manager) dispatchDelegationRecall(handleKey string, lock *UnifiedLock) {
-	lm.mu.RLock()
-	callbacks := make([]BreakCallbacks, len(lm.breakCallbacks))
-	copy(callbacks, lm.breakCallbacks)
-	lm.mu.RUnlock()
-
-	if len(callbacks) == 0 {
-		logger.Debug("delegation recall with no callbacks registered",
-			"handleKey", handleKey,
-			"delegationID", lock.Delegation.DelegationID)
-		return
-	}
-
-	for _, cb := range callbacks {
-		cb.OnDelegationRecall(handleKey, lock)
-	}
-}
-
-// dispatchOpLockBreak notifies all registered break callbacks about an oplock break.
-func (lm *Manager) dispatchOpLockBreak(handleKey string, lock *UnifiedLock, breakToState uint32) {
-	lm.mu.RLock()
-	callbacks := make([]BreakCallbacks, len(lm.breakCallbacks))
-	copy(callbacks, lm.breakCallbacks)
-	lm.mu.RUnlock()
-
-	if len(callbacks) == 0 {
-		logger.Debug("oplock break with no callbacks registered",
-			"handleKey", handleKey,
-			"owner", lock.Owner.OwnerID,
-			"breakToState", LeaseStateToString(breakToState))
-		return
-	}
-
-	for _, cb := range callbacks {
-		cb.OnOpLockBreak(handleKey, lock, breakToState)
-	}
-}
-
-// ============================================================================
-// Grace Period Delegation
-// ============================================================================
-
-// EnterGracePeriod transitions to grace period state.
-// If no grace period manager is configured, this is a no-op.
-func (lm *Manager) EnterGracePeriod(expectedClients []string) {
-	if lm.gracePeriod != nil {
-		lm.gracePeriod.EnterGracePeriod(expectedClients)
-	}
-}
-
-// ExitGracePeriod manually exits the grace period.
-// If no grace period manager is configured, this is a no-op.
-func (lm *Manager) ExitGracePeriod() {
-	if lm.gracePeriod != nil {
-		lm.gracePeriod.ExitGracePeriod()
-	}
-}
-
-// AbortGracePeriod cancels a pending grace timer WITHOUT firing the onGraceEnd
-// callback. Used to discard a manager that lost a registration race: its
-// orphaned timer must never run, because onGraceEnd sweeps the shared lock
-// store (RemoveClientLocks) and ends the surviving NFSv4 grace machine — both
-// would corrupt the published manager's state. If no grace period manager is
-// configured, this is a no-op.
-func (lm *Manager) AbortGracePeriod() {
-	if lm.gracePeriod != nil {
-		lm.gracePeriod.Close()
-	}
-}
-
-// IsOperationAllowed checks if a lock operation is allowed in the current state.
-// If no grace period manager is configured, all operations are allowed.
-func (lm *Manager) IsOperationAllowed(op Operation) (bool, error) {
-	if lm.gracePeriod != nil {
-		return lm.gracePeriod.IsOperationAllowed(op)
-	}
-	return true, nil
-}
-
-// MarkReclaimed records that a client has reclaimed their locks.
-// If no grace period manager is configured, this is a no-op.
-func (lm *Manager) MarkReclaimed(clientID string) {
-	if lm.gracePeriod != nil {
-		lm.gracePeriod.MarkReclaimed(clientID)
-	}
-}
-
-// IsInGracePeriod returns true if grace period is currently active.
-func (lm *Manager) IsInGracePeriod() bool {
-	if lm.gracePeriod != nil {
-		return lm.gracePeriod.GetState() == GraceStateActive
-	}
-	return false
-}
-
-// GetExpectedClients returns the client IDs the grace period is waiting on to
-// reclaim. Returns nil if no grace period manager is configured.
-func (lm *Manager) GetExpectedClients() []string {
-	if lm.gracePeriod != nil {
-		return lm.gracePeriod.GetExpectedClients()
-	}
-	return nil
-}
-
-// GetReclaimedClients returns the client IDs that have reclaimed during the
-// current grace period. Returns nil if no grace period manager is configured.
-func (lm *Manager) GetReclaimedClients() []string {
-	if lm.gracePeriod != nil {
-		return lm.gracePeriod.GetReclaimedClients()
-	}
-	return nil
-}
-
-// ============================================================================
-// Break Callback Registration
-// ============================================================================
-
-// RegisterBreakCallbacks registers typed callbacks for break notifications.
-//
-// Multiple callbacks can be registered (one per protocol adapter).
-// Callbacks are invoked in registration order during break operations.
-func (lm *Manager) RegisterBreakCallbacks(callbacks BreakCallbacks) {
-	lm.mu.Lock()
-	defer lm.mu.Unlock()
-	lm.breakCallbacks = append(lm.breakCallbacks, callbacks)
-}
-
-// ============================================================================
-// Connection/Cleanup Operations
-// ============================================================================
-
-// RemoveAllLocks removes all locks (legacy, unified, and delegations) for a file.
-//
-// The persisted bulk-delete runs synchronously under lm.mu (handleKey is the
-// FileID used when persisting): keeping it ordered with single-record PutLock/
-// DeleteLock prevents a concurrent acquire on the same file from racing this
-// delete to the store and leaving an orphaned record behind (R3-1 class).
-func (lm *Manager) RemoveAllLocks(handleKey string) {
-	lm.mu.Lock()
-	defer lm.mu.Unlock()
-	delete(lm.locks, handleKey)
-	old := lm.unifiedLocks[handleKey]
-	delete(lm.unifiedLocks, handleKey)
-	lm.reindexHandleLocked(handleKey, old)
-	delete(lm.breakWaitChans, handleKey)
-
-	if lm.lockStore != nil {
-		ctx, cancel := withPersistTimeout()
-		defer cancel()
-		if _, err := lm.lockStore.DeleteLocksByFile(ctx, handleKey); err != nil {
-			logger.Error("RemoveAllLocks: failed to delete persisted locks", "handleKey", handleKey, "error", err)
-		}
-	}
-}
-
-// RemoveClientLocks removes all unified locks held by a specific client. The
-// persisted bulk-delete runs synchronously under lm.mu for the same ordering
-// reason as RemoveAllLocks.
-//
-// clientHandleIndex (clientID -> set of handleKeys, see indexes.go) bounds the
-// work to the buckets the client actually appears in instead of scanning every
-// entry in unifiedLocks under the global mutex.
-func (lm *Manager) RemoveClientLocks(clientID string) {
-	lm.mu.Lock()
-	defer lm.mu.Unlock()
-
-	// Snapshot the affected handleKeys before mutating: reindexHandleLocked
-	// updates clientHandleIndex underneath us, so iterating the map directly
-	// would mutate it mid-range.
-	var handleKeys []string
-	for handleKey := range lm.clientHandleIndex[clientID] {
-		handleKeys = append(handleKeys, handleKey)
-	}
-
-	for _, handleKey := range handleKeys {
-		locks := lm.unifiedLocks[handleKey]
-		var remaining []*UnifiedLock
-		for _, lock := range locks {
-			if lock.Owner.ClientID != clientID {
-				remaining = append(remaining, lock)
-			}
-		}
-		if len(remaining) == 0 {
-			delete(lm.unifiedLocks, handleKey)
-		} else {
-			lm.unifiedLocks[handleKey] = remaining
-		}
-		lm.reindexHandleLocked(handleKey, locks)
-	}
-
-	if lm.lockStore != nil {
-		ctx, cancel := withPersistTimeout()
-		defer cancel()
-		if _, err := lm.lockStore.DeleteLocksByClient(ctx, clientID); err != nil {
-			logger.Error("RemoveClientLocks: failed to delete persisted locks", "clientID", clientID, "error", err)
-		}
-	}
-}
-
-// ReleaseByOwnerPrefix removes every unified lock whose Owner.OwnerID begins
-// with the given prefix and returns the number of locks released.
-//
-// This is used for NSM crash cleanup: when a client crashes, NLM must release
-// all locks it held. NLM owner IDs are formatted as
-// "nlm:{caller_name}:{svid}:{oh_hex}", so passing "nlm:{caller_name}:" releases
-// every byte-range lock the crashed client held across all files in this share
-// without touching locks from other protocols (SMB/NFSv4 use different prefixes)
-// or other NLM clients. The trailing ":" in the caller-supplied prefix prevents
-// "nlm:client1:" from matching "nlm:client10:".
-//
-// The persisted bulk-delete runs synchronously under lm.mu for the same
-// ordering reason as RemoveClientLocks. Calling with a prefix that matches no
-// locks is safe and returns 0.
-//
-// Unlike RemoveClientLocks, this intentionally keeps the full unifiedLocks
-// scan: the predicate is a prefix match on Owner.OwnerID, and there is no
-// reverse index keyed by OwnerID prefix (an exact-OwnerID index could not
-// answer a prefix query without either materializing every prefix or scanning
-// the index keys anyway, so it would not bound the work). NSM crash cleanup is
-// rare and not on the steady-state lease/lock hot path, so the scan is
-// acceptable; the index maintenance below still keeps clientHandleIndex /
-// leaseKeyIndex consistent for the records this removes.
-func (lm *Manager) ReleaseByOwnerPrefix(prefix string) int {
-	lm.mu.Lock()
-	defer lm.mu.Unlock()
-
-	released := 0
-	for handleKey, locks := range lm.unifiedLocks {
-		var remaining []*UnifiedLock
-		mutated := false
-		for _, lock := range locks {
-			if strings.HasPrefix(lock.Owner.OwnerID, prefix) {
-				lm.deleteUnifiedLockLocked(lock)
-				released++
-				mutated = true
-				continue
-			}
-			remaining = append(remaining, lock)
-		}
-		if !mutated {
-			continue
-		}
-		if len(remaining) == 0 {
-			delete(lm.unifiedLocks, handleKey)
-		} else {
-			lm.unifiedLocks[handleKey] = remaining
-		}
-		lm.reindexHandleLocked(handleKey, locks)
-	}
-
-	return released
-}
-
-// GetStats returns current lock manager statistics.
-func (lm *Manager) GetStats() ManagerStats {
-	lm.mu.RLock()
-	defer lm.mu.RUnlock()
-
-	totalLegacy := 0
-	for _, locks := range lm.locks {
-		totalLegacy += len(locks)
-	}
-
-	totalUnified := 0
-	for _, locks := range lm.unifiedLocks {
-		totalUnified += len(locks)
-	}
-
-	fileSet := make(map[string]struct{})
-	for key := range lm.locks {
-		fileSet[key] = struct{}{}
-	}
-	for key := range lm.unifiedLocks {
-		fileSet[key] = struct{}{}
-	}
-
-	return ManagerStats{
-		TotalLegacyLocks:   totalLegacy,
-		TotalUnifiedLocks:  totalUnified,
-		TotalFiles:         len(fileSet),
-		BreakCallbackCount: len(lm.breakCallbacks),
-		GracePeriodActive:  lm.gracePeriod != nil && lm.gracePeriod.GetState() == GraceStateActive,
-	}
 }
